// model/bloodRequest.tsx
export interface BloodRequest {
  id?: string;
  name?: string;
  bloodGroup?: string;
  hospital?: string;
  location?: string;
  date?: string;
  deadline?: string;
  phone?: string;
  reason?: string;
  status?: string;
  userId?: string;
  createdAt?: string;
  updatedAt?: string;
  age?: number;
}